Support zeroed out TAMM References
By defining a FRAM Fishery/Stock in the TAMM reference tables without a
worksheet name, column and row, the original FRAM Value will be replaced
with a zero.

Note: I slightly reformated the "GetTammValues" function so that the
order of the variables works better with pipes and updated the function
documentation to a newer format.